+++ packages/sdk/CHANGELOG.md
@@ -7,6 +7,10 @@
 - `toString`, `toJSON`, and `fromJSON` to all concordium domain types.
 - Deprecated types and functions related to typed JSON serialization and deserialization.
 
+### Fixed
+
+- Changed the return type of `TransactionExpiry.toJSON` from `number` to `bigint`, since the former could throw an error.
+
 ## 7.3.2
 
 ### Added